Fixed to have no errors -- no guarantee that old classes will work as intended, though.

+++ src/org/srg/scpp_im/game/SAAHomogeneousGame.java
@@ -241,7 +241,7 @@ public class SAAHomogeneousGame extends GameSetting implements Register {
 	private void run()
 	{
 		//ArrayList<int[]> bids = new ArrayList<int[]>();
-		int[][] bids = new int[NUM_AGENT][NUM_GOODS];
+		double[][] bids = new double[NUM_AGENT][NUM_GOODS];
 		// testing a single SAA
 		boolean isQuiescent = false;
 
@@ -249,13 +249,13 @@ public class SAAHomogeneousGame extends GameSetting implements Register {
 		while (!isQuiescent)
 		{
 			// Supply current information state to users
-			int[] currentBids = state.getCurrentBidPrice();
+			double[] currentBids = state.getCurrentBidPrice();
 			int[] currentWinning = state.getCurrentBidWinning();
 
 			for (int i=0;i<NUM_AGENT;i++)
 			{
 				Strategy s = strategies.get(i);
-				int[] newbid = s.bid(state);
+				double[] newbid = s.bid(state);
 				
 				//System.out.print("Agent " + s.getIndex() + " bids: ");
 				for (int j=0;j<NUM_GOODS;j++)
@@ -268,7 +268,7 @@ public class SAAHomogeneousGame extends GameSetting implements Register {
 			isQuiescent = true;
 			for (int i=0;i<NUM_GOODS;i++)
 			{
-				int max = currentBids[i];
+				double max = currentBids[i];
 				//int agentWon = 0;
 				for (int j=0; j<NUM_AGENT;j++)
 				{
@@ -326,7 +326,7 @@ public class SAAHomogeneousGame extends GameSetting implements Register {
 			s.addObservation(state);
 		}
 		
-		int[] currentBids = state.getCurrentBidPrice();
+		double[] currentBids = state.getCurrentBidPrice();
 		int[] currentWinning = state.getCurrentBidWinning();
 		for (int i=0;i<NUM_GOODS;i++)
 		{
