+++ src/org/srg/scpp_im/strategy/OSSCDP_BidEvaluatorSMU_DR.java
@@ -26,9 +26,9 @@ public class OSSCDP_BidEvaluatorSMU_DR extends
 		return "OSSCDP_StraightMU";
 	}
 
-	public int[] bid(InformationState s)
+	public double[] bid(InformationState s)
 	{
-		int[] newBid = new int[NUM_GOODS];
+		double[] newBid = new double[NUM_GOODS];
 		int[] singleGoodValue = new int[NUM_GOODS];
 		int[] priceToBid = new int[NUM_GOODS];
 		demand = new double[NUM_GOODS];
